deprecate infra validate, and cleaned up dependencies
