+++ src/commands/infra/index.ts
@@ -1,10 +1,9 @@
 import { Command } from "../command";
 import { commandDecorator as generateCommandDecorator } from "./generate";
 import { commandDecorator as scaffoldCommandDecorator } from "./scaffold";
-import { commandDecorator as validateCommandDecorator } from "./validate";
 
 export const infraCommand = Command(
   "infra",
   "Manage and modify your Bedrock infrastructure.",
-  [generateCommandDecorator, scaffoldCommandDecorator, validateCommandDecorator]
+  [generateCommandDecorator, scaffoldCommandDecorator]
 );
